Update Swift client to new style
